Create Carousel Component and Data Model
Since our app is a carousel, carousel.tsx handles all the state. The
index.tsx entry point passes our dataImage to the carousel component,
the component initializes ImageData to manage the entries and image
counts.

SetState is called when the carousel buttons are incremented or
decremented, and the currentImage and image count is updated.